Reworked ScopedContextData Tests (#539)
